requirements/others.txt
# Tensorboard
tensorboard

# Jypyter
jupyterlab

# Utils
easydict
tqdm
seaborn
torchmetrics